Bot class restructure
